chore: remove card and carnumber components
